// src/enums/Modules.ts
/**
 * Enum representing module parameter list.
 */
enum Modules {
  BANK_GROUP = 'bankGroup',
  BANK_TYPE = 'bankType',
  BUCKET_PROCESS = 'bucketProcess',
  COLLECTIBILITY = 'collectibility',
  COUNTRY = 'country',
  CURRENCY = 'currency',
  DEBTOR_TYPE = 'debtorType',
  DISTRICT = 'districtIdn',
  FINANCING_FACILITY_OTHER_BANK = 'financingFacilityOtherBankNameList',
  FINANCING_SEGMENT = 'financingSegment',
  FINANCING_TYPE = 'financingType',
  GOVERMENT_MANDATE = 'governmentMandate',
  GOLONGAN_KREDIT = 'golonganKredit',
  GRADE_CAP = 'gradeCap',
  GRADE_REPORT = 'gradeReport',
  IDENTIFY_RISKS = 'typeOfRiskDH',
  INSTITUTION_TYPE = 'institutionType',
  JOB_POSITION = 'jobPosition',
  LOAN_TYPE = 'loanType',
  MAPPING_ORDER_TYPE = 'mappingOrderTypePK',
  ORDER_TYPE = 'orderType',
  ORDER_TYPE_EXISTING = 'orderTypeExisting',
  ORIENTASI_PENGGUNAAN = 'orientasiPenggunaan',
  PROFILE_RISKS = 'typeOfRiskDEPI',
  RATING = 'rating',
  SEARCH_BY_FACILITY_EXISTING = 'searchByFacilityExisting',
  SEARCH_INQUIRY_DATA = 'searchByInquiryData',
  SORT_BY_FACILITY_EXISTING_DATA = 'sortByFacilityExistingData',
  SORT_INQUIRY_DATA = 'sortByInquiryData',
  TYPE_OF_USE = 'typeOfUse',
  TYPE_PROCESS = 'typeProcess',
  TYPE_SUBMISSION = 'typeSubmission',
  ID_DOC_TYPE = 'idDocType',
  APPLICATION_CATEGORY = 'apApplicationCategory'
};

export default Modules;