// src/services/openapi/loan-service/api.ts
export interface ProjectDto {
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    value?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    curValue?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    sector?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    sectorLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    district?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    districtLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    city?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    cityLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    province?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    provinceLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    curExchangeRate?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    exchangeRate?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    curValueInIdr?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    valueInIdr?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    debtorId?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    createdDate?: string;
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    createdBy?: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    modifiedDate?: string;
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    modifiedBy?: number;
}